# docs/swagger.yaml
definitions:
...
  requests.CreateCampaignRequest:
    properties:
      contract_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      description:
        example: A campaign for the summer sale.
        maxLength: 1000
        type: string
      end_date:
        example: "2023-08-31T23:59:59Z"
        type: string
      milestones:
        items:
          $ref: '#/definitions/requests.CreateMilestoneCampaignRequest'
        type: array
      name:
        example: Summer Sale Campaign
        maxLength: 255
        minLength: 3
        type: string
      start_date:
        example: "2023-06-01T00:00:00Z"
        type: string
      type:
        enum:
        - ADVERTISING
        - AFFILIATE
        - AMBASSADOR
        - COPRODUCE
        example: ADVERTISING
        type: string
    required:
    - contract_id
    - end_date
    - name
    - start_date
    - type
    type: object
...
  responses.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
...
  responses.CampaignDetailsResponse:
    properties:
      budget:
        example: 100000000
        type: number
      contract_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      contract_number:
        example: contract_20251017_AD
        type: string
      contract_title:
        example: Q2 Marketing Contract
        type: string
      created_at:
        example: "2023-06-01 00:00:00"
        type: string
      description:
        example: A campaign for the summer sale.
        type: string
      end_date:
        example: "2023-08-31 23:59:59"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      milestones:
        items:
          $ref: '#/definitions/responses.CampaignMilestoneInfo'
        type: array
      name:
        example: Summer Sale Campaign
        type: string
      number_of_tasks:
        example: 25
        type: integer
      percentage_completed:
        example: 60.5
        type: number
      start_date:
        example: "2023-06-01 00:00:00"
        type: string
      status:
        example: RUNNING
        type: string
      type:
        example: ADVERTISING
        type: string
      updated_at:
        example: "2023-06-15 12:00:00"
        type: string
    type: object
  responses.CampaignInfoPaginationResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/responses.CampaignInfoResponse'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/responses.Pagination'
      status:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
...
  responses.CampaignSuggestionResponse:
    properties:
      contract_id:
        type: string
      contract_type:
        type: string
      suggested_campaign:
        $ref: '#/definitions/responses.SuggestedCampaign'
    type: object
...
  responses.Pagination:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  responses.SuggestedCampaign:
    properties:
      milestones:
        items:
          $ref: '#/definitions/responses.SuggestedMilestone'
        type: array
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the core backend service for SEP490 project with authentication,
    user management, and business features.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SEP490 Core Backend API
  version: "1.0"
paths:
...
  /api/v1/campaigns:
    get:
      consumes:
      - application/json
      description: Get paginated list of campaigns with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search keywords for campaign name
        in: query
        name: keywords
        type: string
      - description: Filter by campaign status
        enum:
        - ACTIVE
        - INACTIVE
        in: query
        name: status
        type: string
      - description: Filter by campaign type
        enum:
        - ADVERTISING
        - AFFILIATE
        - BRAND_AMBASSADOR
        - CO_PRODUCING
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Campaigns retrieved successfully
          schema:
            $ref: '#/definitions/responses.CampaignInfoPaginationResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Campaigns List by filter
      tags:
      - Campaigns
    post:
      consumes:
      - application/json
      description: Create a new campaign based on the provided contract details.
      parameters:
      - description: Campaign creation data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.CreateCampaignRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Campaign created successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/responses.CampaignDetailsResponse'
              type: object
        "400":
          description: Invalid request or validation error
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Create Campaign from Contract
      tags:
      - Campaigns
  /api/v1/campaigns/{campaign_id}/suggest:
    get:
      consumes:
      - application/json
      description: Generate campaign suggestions based on contract deliverables. Only
        ACTIVE contracts can be used for suggestions.
      parameters:
      - description: Campaign ID
        format: uuid
        in: path
        name: campaign_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Campaign suggestion generated successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/responses.CampaignSuggestionResponse'
              type: object
        "400":
          description: Invalid request or validation error
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "403":
          description: Forbidden - Requires ADMIN or SALES_STAFF role
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Contract not found
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "409":
          description: Contract is not ACTIVE or has no deliverables
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Suggest Campaign from Contract
      tags:
      - Campaigns
  /api/v1/campaigns/brand/{brand_id}:
    get:
      consumes:
      - application/json
      description: Get paginated list of campaigns with optional filters
      parameters:
      - description: Brand ID
        format: uuid
        in: path
        name: brand_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Campaigns retrieved successfully
          schema:
            $ref: '#/definitions/responses.CampaignInfoPaginationResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Campaigns Info by Brand ID
      tags:
      - Campaigns
  /api/v1/campaigns/brand/profile:
    get:
      consumes:
      - application/json
      description: Get a list of campaigns associated with the authenticated brand
        user.
      produces:
      - application/json
      responses:
        "200":
          description: Campaigns retrieved successfully
          schema:
            $ref: '#/definitions/responses.CampaignInfoPaginationResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Campaigns Info by authenticated brand user
      tags:
      - Campaigns
  /api/v1/campaigns/contract/{contract_id}:
    get:
      consumes:
      - application/json
      description: Get basic information about a campaign by its contract ID.
      parameters:
      - description: Contract ID
        format: uuid
        in: path
        name: contract_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Campaign info retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/responses.CampaignInfoResponse'
              type: object
        "400":
          description: Invalid contract ID
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Campaign not found
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Campaign Info by Contract ID
      tags:
      - Campaigns
  /api/v1/campaigns/contract/{contract_id}/details:
    get:
      consumes:
      - application/json
      description: Get detailed information about a campaign by its contract ID including
        milestones and number of tasks.
      parameters:
      - description: Contract ID
        format: uuid
        in: path
        name: contract_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Campaign details retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/responses.CampaignDetailsResponse'
              type: object
        "400":
          description: Invalid contract ID
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Campaign not found
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Campaign Details by Contract ID
      tags:
      - Campaigns
  /api/v1/campaigns/id/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a campaign by ID
      parameters:
      - description: Campaign ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Campaign deleted successfully
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "400":
          description: Invalid campaign ID
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Campaign not found
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete Campaign
      tags:
      - Campaigns
    get:
      consumes:
      - application/json
      description: Get basic information about a campaign by its ID.
      parameters:
      - description: Campaign ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Campaign info retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/responses.CampaignInfoResponse'
              type: object
        "400":
          description: Invalid campaign ID
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Campaign not found
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Campaign Info by ID
      tags:
      - Campaigns
  /api/v1/campaigns/id/{id}/details:
    get:
      consumes:
      - application/json
      description: Get detailed information about a campaign by its ID including milestones
        and number of tasks.
      parameters:
      - description: Campaign ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Campaign details retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/responses.CampaignDetailsResponse'
              type: object
        "400":
          description: Invalid campaign ID
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Campaign not found
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Campaign Details by ID
      tags:
      - Campaigns
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"